requirements.txt
numpy==1.15.0
py-cpuinfo==4.0.0

